+++ main/bundle_of_joy/lib/appointmentMother/appointmentMother_add_2.dart
@@ -1,0 +1,221 @@
+import "package:flutter/material.dart";
+import "appointmentMother_add_1.dart";
+import "appointmentMother_add_3.dart";
+
+class AppointmentMotherAdd2 extends StatefulWidget {
+  final String name;
+
+  AppointmentMotherAdd2({this.name});
+
+  @override
+  _AppointmentMotherAdd2State createState() => _AppointmentMotherAdd2State(name);
+}
+
+class _AppointmentMotherAdd2State extends State<AppointmentMotherAdd2> {
+  // VARIABLES
+  final String hospitalName;
+  DateTime pickedDate;
+
+  // MAKE THE DEFAULT DATE TODAY
+  @override
+  void initState() {
+    super.initState();
+    pickedDate = DateTime.now();
+  }
+
+  _AppointmentMotherAdd2State(this.hospitalName);
+
+  // BUILD THE WIDGET
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      // APP BAR
+      appBar: AppBar(
+        title: Text(
+          "Appointment Management",
+          style: TextStyle(
+            fontFamily: 'Comfortaa',
+            fontWeight: FontWeight.bold,
+            fontSize: MediaQuery.of(context).size.width * 0.05,
+            color: Colors.black,
+          ),
+        ),
+
+        automaticallyImplyLeading: false, // CENTER THE TEXT
+        backgroundColor: Color(0xFFFCFFD5),
+        centerTitle: true,
+      ),
+
+      // BODY
+      body: Column(
+        children: <Widget>[
+          Container(
+              // color: Colors.lightBlue,
+              height: MediaQuery.of(context).size.height * 0.58,
+              width: MediaQuery.of(context).size.width,
+              margin: EdgeInsets.only(bottom: MediaQuery.of(context).size.height * 0.03),
+              child: SingleChildScrollView(
+                child: Column(
+                  children: [
+                    SizedBox(height: MediaQuery.of(context).size.height * 0.04),
+                    Container(
+                      height: MediaQuery.of(context).size.height * 0.04,
+                      width: MediaQuery.of(context).size.width * 0.8,
+                      child: Text(
+                        hospitalName,
+                        style: TextStyle(
+                          fontFamily: 'Comfortaa',
+                          fontWeight: FontWeight.bold,
+                          fontSize: MediaQuery.of(context).size.height * 0.03,
+                          color: Colors.black,
+                        ),
+                        textAlign: TextAlign.center,
+                        overflow: TextOverflow.ellipsis,
+                        maxLines: 1,
+                        softWrap: true,
+                      ),
+                    ),
+                    Container(
+                      margin: EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.03, bottom: MediaQuery.of(context).size.height * 0.06),
+                      child: Text(
+                        "${pickedDate.day} - ${pickedDate.month} - ${pickedDate.year}",
+                        style: TextStyle(
+                          fontFamily: 'Comfortaa',
+                          fontWeight: FontWeight.bold,
+                          fontSize: MediaQuery.of(context).size.height * 0.028,
+                          color: Colors.black,
+                        ),
+                      ),
+                    ),
+                    Container(
+                      height: MediaQuery.of(context).size.height * 0.30,
+                      width: MediaQuery.of(context).size.width * 0.75,
+                      decoration: myBoxDecoration(),
+                      child: Column(
+                        children: [
+                          InkWell(
+                            child: Container(
+                              margin: EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.01, bottom: MediaQuery.of(context).size.height * 0.01),
+                              child: Image.asset(
+                                "assets/icons/calendar.png",
+                                height: MediaQuery.of(context).size.height * 0.2,
+                              ),
+                            ),
+                            onTap: _pickDate,
+                          ),
+                          Text(
+                            "Pick A Date",
+                            style: TextStyle(
+                              fontFamily: 'Comfortaa',
+                              fontWeight: FontWeight.bold,
+                              fontSize: MediaQuery.of(context).size.height * 0.035,
+                              color: Colors.black,
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              )),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              InkWell(
+                child: Container(
+                  width: MediaQuery.of(context).size.width * 0.3,
+                  height: MediaQuery.of(context).size.height * 0.06,
+                  decoration: myBoxDecoration2(),
+                  child: Center(
+                    child: Text(
+                      "Back",
+                      style: TextStyle(
+                        fontFamily: 'Comfortaa',
+                        fontWeight: FontWeight.bold,
+                        fontSize: MediaQuery.of(context).size.height * 0.025,
+                        color: Colors.black,
+                      ),
+                    ),
+                  ),
+                ),
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => AppointmentMotherAdd1()),
+                  );
+                },
+              ),
+              SizedBox(width: MediaQuery.of(context).size.width * 0.05),
+              InkWell(
+                child: Container(
+                  width: MediaQuery.of(context).size.width * 0.3,
+                  height: MediaQuery.of(context).size.height * 0.06,
+                  decoration: myBoxDecoration2(),
+                  child: Center(
+                    child: Text(
+                      "Next",
+                      style: TextStyle(
+                        fontFamily: 'Comfortaa',
+                        fontWeight: FontWeight.bold,
+                        fontSize: MediaQuery.of(context).size.height * 0.025,
+                        color: Colors.black,
+                      ),
+                    ),
+                  ),
+                ),
+                onTap: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(builder: (context) => AppointmentMotherAdd3(name: hospitalName, date: pickedDate)),
+                  );
+                },
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+
+  _pickDate() async {
+    DateTime date = await showDatePicker(
+      context: context,
+      firstDate: DateTime(DateTime.now().year - 5),
+      lastDate: DateTime(DateTime.now().year + 5),
+      initialDate: pickedDate,
+      builder: (BuildContext context, Widget child) {
+        return Theme(
+          data: ThemeData(primarySwatch: Colors.pink, splashColor: Colors.green),
+          child: child,
+        );
+      },
+    );
+
+    if (date != null) {
+      setState(() {
+        pickedDate = date;
+      });
+    }
+  }
+
+  BoxDecoration myBoxDecoration() {
+    return BoxDecoration(
+      color: Color(0xFFFCFFD5),
+      border: Border.all(
+        color: Colors.black,
+        width: 5.0,
+      ),
+    );
+  }
+
+  BoxDecoration myBoxDecoration2() {
+    return BoxDecoration(
+      color: Color(0xFFFCFFD5),
+      border: Border.all(
+        color: Colors.black,
+        width: 2.0,
+      ),
+      borderRadius: BorderRadius.all(Radius.circular(10.0)),
+    );
+  }
+}
